requirements.txt
streamlit
pandas
plotly
faiss-cpu
sentence-transformers
numpy
requests
torch